requirements.txt
-i https://pypi.org/simple
certifi==2024.6.2 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
contourpy==1.2.1 ; python_version >= '3.9'
cycler==0.12.1 ; python_version >= '3.8'
fonttools==4.53.0 ; python_version >= '3.8'
idna==3.7 ; python_version >= '3.5'
kiwisolver==1.4.5 ; python_version >= '3.7'
matplotlib==3.9.0 ; python_version >= '3.9'
numpy==2.0.0 ; python_version == '3.11'
packaging==24.1 ; python_version >= '3.8'
pandas==2.2.2
pillow==10.3.0 ; python_version >= '3.8'
plotly==5.22.0
pyparsing==3.1.2 ; python_full_version >= '3.6.8'
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
so4t-api==0.1.5
tenacity==8.4.1 ; python_version >= '3.8'
tzdata==2024.1 ; python_version >= '2'
urllib3==2.2.2 ; python_version >= '3.8'
wordcloud==1.9.3
